Enable extra warnings of gcc and fix them.

See merge request c39gao/cs452-kernel!4

+++ include/kern/ipc.h
@@ -1,19 +1,22 @@
 #ifndef __KERN_IPC_H___
 #define __KERN_IPC_H___
 
+#include <stddef.h>
+#include <stdint.h>
+
 typedef struct {
     char *array;
-    unsigned int len;
+    size_t len;
 } Message;
 
 // ipc_send() sends message(msg) to recvtid, and waits for a reply message to be written to reply
-void ipc_send(int tid, int recvtid, char *msg, int msglen, char *reply, int rplen);
+void ipc_send(int tid, int recvtid, char *msg, size_t msglen, char *reply, size_t rplen);
 
 // ipc_receive() copies a sender's message to msg and the sender's tid to sendtid.
-void ipc_receive(int tid, int *sendtid, char *msg, int msglen);
+void ipc_receive(int tid, int *sendtid, char *msg, size_t msglen);
 
 // ipc_reply() copies the reply message from tid Task to replytid Task.
-void ipc_reply(int tid, int replytid, char *reply, int rplen);
+void ipc_reply(int tid, int replytid, char *reply, size_t rplen);
 
 // ipc_cleanup() clears the send queue by unblock the tasks, and return an error code.
 // for incomplete transaction
